requirements.txt
# Core FastAPI and server requirements
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.20

# Database requirements
asyncpg==0.30.0
pgvector==0.4.1

# LangChain ecosystem for AI agents
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9

# Google AI API requirements
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0

# Environment and configuration
python-dotenv==1.1.1

# Core dependencies (automatically pulled by above packages)
# These are typically installed as dependencies of the above packages:
# - aiohttp, aiosignal, anyio, attrs, certifi, charset-normalizer
# - click, frozenlist, grpcio, grpcio-status, h11, httpcore, httptools
# - httpx, idna, jsonpatch, jsonpointer, jsonschema, multidict
# - packaging, proto-plus, protobuf, pyasn1, pyasn1_modules
# - pydantic-core, requests, sniffio, starlette, tenacity
# - typing_extensions, urllib3, yarl
